+++ Lead-Client-Source/GameLib/RaceMotionData.cpp
@@ -337,9 +337,9 @@ bool CRaceMotionData::LoadMotionData(const char * c_szFileName)
 			return false;
 		}
 
-		TPixelPosition pos(atof(pTokenVector->at(0).c_str()),
-						   atof(pTokenVector->at(1).c_str()),
-						   atof(pTokenVector->at(2).c_str()));
+		TPixelPosition pos(static_cast<float>(atof(pTokenVector->at(0).c_str())),
+						   static_cast<float>(atof(pTokenVector->at(1).c_str())),
+						   static_cast<float>(atof(pTokenVector->at(2).c_str())));
 
 		SetAccumulationPosition(pos);
 	}
@@ -447,7 +447,7 @@ bool CRaceMotionData::LoadMotionData(const char * c_szFileName)
 				if (!rkTextFileLoader.GetTokenFloat("startingtime", &m_MotionEventDataVector[j]->fStartingTime))
 					return false;
 
-				m_MotionEventDataVector[j]->dwFrame = (m_MotionEventDataVector[j]->fStartingTime / c_fFrameTime);
+				m_MotionEventDataVector[j]->dwFrame = static_cast<DWORD>(m_MotionEventDataVector[j]->fStartingTime / c_fFrameTime);
 
 				rkTextFileLoader.SetParentNode();
 			}
